+++ sdk/monitor/azure-monitor-opentelemetry-exporter/CHANGELOG.md
@@ -8,6 +8,8 @@
     ([#25752](https://github.com/Azure/azure-sdk-for-python/pull/25752))
 - Implement all network statsbeat
     ([#25845](https://github.com/Azure/azure-sdk-for-python/pull/25845))
+- Implement attach statsbeat
+    ([#25845](https://github.com/Azure/azure-sdk-for-python/pull/25845))
 
 ### Breaking Changes